pyproject.toml
[project]
name = "event-scheduler"
version = "0.1.0"
description = "Event Scheduler Application built with FastAPI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.2",
    "authlib>=1.3",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.1",
    "openai>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
    "black>=24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
